requirements.txt
allure-pytest==2.8.22
allure-python-commons==2.8.22
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
docopt==0.6.2
EasyProcess==0.3
idna==2.10
iniconfig==1.1.1
packaging==20.4
pipreqs==0.4.10
pluggy==0.13.1
py==1.9.0
PyAutoIt==0.6.0
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==6.1.2
pytest-html==3.0.0
pytest-metadata==1.11.0
requests==2.25.0
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.2
urllib3==1.26.2
yarg==0.1.9
